requirements.txt
PyGitHub
pandas
requests
radon
pylint
textblob
python-dotenv
streamlit
flake8
ollama
